Refactoring collisions functionality (#1769)
