+++ changelog/1678.trivial.1.rst
@@ -1,2 +1,2 @@
 Re-enabled value testing for the ``quantum`` keyword argument in
-`~plasmapy.formulary.collisions.coupling_parameter`.
+`~plasmapy.formulary.collisions.dimensionless.coupling_parameter`.
