+++ changelog/1703.breaking.2.rst
@@ -1,1 +1,4 @@
-Deprecated ``plasmapy.formulary.collisions.fundamental_ion_collision_freq``. The `~plasmapy.formulary.collisions.MaxwellianCollisionFrequencies.Maxwellian_avg_ii_collision_freq` attribute of `~plasmapy.formulary.collisions.MaxwellianCollisionFrequencies` should be used instead.
+Deprecated ``plasmapy.formulary.collisions.fundamental_ion_collision_freq``.
+The `~plasmapy.formulary.collisions.frequencies.MaxwellianCollisionFrequencies.Maxwellian_avg_ii_collision_freq`
+attribute of `~plasmapy.formulary.collisions.frequencies.MaxwellianCollisionFrequencies`
+should be used instead.
