+++ docs/api_static/plasmapy.formulary.collisions.coulomb.rst
@@ -1,0 +1,8 @@
+:orphan:
+
+`plasmapy.formulary.collisions.coulomb`
+=======================================
+
+.. currentmodule:: plasmapy.formulary.collisions.coulomb
+
+.. automodapi:: plasmapy.formulary.collisions.coulomb
