+++ docs/api_static/plasmapy.formulary.collisions.dimensionless.rst
@@ -1,0 +1,8 @@
+:orphan:
+
+`plasmapy.formulary.collisions.dimensionless`
+=============================================
+
+.. currentmodule:: plasmapy.formulary.collisions.dimensionless
+
+.. automodapi:: plasmapy.formulary.collisions.dimensionless
