+++ docs/api_static/plasmapy.formulary.collisions.frequencies.rst
@@ -1,0 +1,8 @@
+:orphan:
+
+`plasmapy.formulary.collisions.frequencies`
+===========================================
+
+.. currentmodule:: plasmapy.formulary.collisions.frequencies
+
+.. automodapi:: plasmapy.formulary.collisions.frequencies
