+++ docs/api_static/plasmapy.formulary.collisions.lengths.rst
@@ -1,0 +1,8 @@
+:orphan:
+
+`plasmapy.formulary.collisions.lengths`
+=======================================
+
+.. currentmodule:: plasmapy.formulary.collisions.lengths
+
+.. automodapi:: plasmapy.formulary.collisions.lengths
