+++ docs/api_static/plasmapy.formulary.collisions.misc.rst
@@ -1,0 +1,8 @@
+:orphan:
+
+`plasmapy.formulary.collisions.misc`
+====================================
+
+.. currentmodule:: plasmapy.formulary.collisions.misc
+
+.. automodapi:: plasmapy.formulary.collisions.misc
