+++ docs/whatsnew/0.8.1.rst
@@ -256,7 +256,7 @@ Features
   a given 3D magnetic null point. (`#1554
   <https://github.com/plasmapy/plasmapy/pull/1554>`__)
 - Added support for arbitrarily shaped input arrays to the function
-  `plasmapy.formulary.collisions.impact_parameter`. (`#1604
+  `plasmapy.formulary.collisions.lengths.impact_parameter`. (`#1604
   <https://github.com/plasmapy/plasmapy/pull/1604>`__)
 
 
@@ -311,9 +311,9 @@ Bug Fixes
   allow for arguments ``theta`` and ``k`` to simultaneously be
   arrays. (`#1529 <https://github.com/plasmapy/plasmapy/pull/1529>`__)
 - Fixed the ``Z`` dependence in
-  `~plasmapy.formulary.collisions.fundamental_electron_collision_freq`,
+  `~plasmapy.formulary.collisions.frequencies.fundamental_electron_collision_freq`,
   by replacing ``n_e`` with ``n_i`` while calling
-  `~plasmapy.formulary.collisions.collision_frequency`. (`#1546
+  `~plasmapy.formulary.collisions.frequencies.collision_frequency`. (`#1546
   <https://github.com/plasmapy/plasmapy/pull/1546>`__)
 - Updated the regular expression matching used by
   `~plasmapy.particles.particle_class.Particle` to parse and identify
@@ -389,7 +389,7 @@ Improved Documentation
   |bibliography|. (`#1374
   <https://github.com/plasmapy/plasmapy/pull/1374>`__)
 - Fixed the docstring of
-  `~plasmapy.formulary.collisions.coupling_parameter`. (`#1379
+  `~plasmapy.formulary.collisions.dimensionless.coupling_parameter`. (`#1379
   <https://github.com/plasmapy/plasmapy/pull/1379>`__)
 - Added an example notebook that introduces how to use
   `astropy.units`. (`#1380
@@ -666,7 +666,7 @@ Trivial/Internal Changes
   <https://sourcery.ai/>`__ to `plasmapy.analysis`. (`#1564
   <https://github.com/plasmapy/plasmapy/pull/1564>`__)
 - Removed an extraneous `print` statement from
-  `~plasmapy.formulary.collisions.collision_frequency` that activated
+  `~plasmapy.formulary.collisions.frequencies.collision_frequency` that activated
   when the colliding particles were both electrons. (`#1570
   <https://github.com/plasmapy/plasmapy/pull/1570>`__)
 - Changed the type hints for ``z_mean`` in
