+++ plasmapy/formulary/__init__.py
@@ -25,14 +25,18 @@ from plasmapy.formulary.relativity import *
 from plasmapy.formulary.speeds import *
 
 # auto populate __all__
+import inspect  # isort: skip
+
 for obj_name in list(globals()):
-    if not (obj_name.startswith("__") or obj_name.endswith("__")):
+    if not (
+        obj_name.startswith("__") or obj_name.endswith("__")
+    ) and not inspect.ismodule(globals()[obj_name]):
         __all__.append(obj_name)
 __all__.sort()
 
 # Put non-formulary imports here so that they don't get included in __all__
 
-import contextlib  # isort:skip
+import contextlib  # isort: skip
 
 # auto populate __aliases__ & __lite_funcs__
 for modname in (
@@ -70,4 +74,4 @@ __aliases__ = list(sorted(set(__aliases__)))
 __lite_funcs__ = list(sorted(set(__lite_funcs__)))
 
 # cleanup namespace
-del contextlib, modname, obj, obj_name
+del contextlib, inspect, modname, obj, obj_name
