+++ setup.cfg
@@ -284,7 +284,8 @@ per-file-ignores =
     plasmapy/analysis/nullpoint.py:SIM102,SIM103
     plasmapy/diagnostics/charged_particle_radiography.py:FS003
     plasmapy/formulary/__init__.py:E402,F403
-    plasmapy/formulary/collisions.py:SIM114
+    plasmapy/formulary/collisions/__init__.py:E402,F403
+    plasmapy/formulary/collisions/lengths.py:SIM114
     # The decorators.py ignore is temporary pending #1057
     plasmapy/particles/decorators.py:SIM102
     plasmapy/particles/ionization_state_collection.py:SIM102
